Add IP handling to Wi-Fi Access Point sample Application 'wifiApTest'
The WiFi AP is now able to "automatically" reconfigure itself at each
start. The actual limitations are the AP IP address and the IP range
address which are fixed for the moment.

Patch 1:
Check the file handle validity...

Patch 2:
Typo correction.
Correction related to the checking of the returned system call.

Change-Id: I7741c2fcf5ecb4feda4733fb439f1b6a4757846b
Resolved: #4681