Add IP handling to Wi-Fi Access Point sample Application 'wifiWebAPPt'
The WiFi AP is a 3G gateway and can now be configured,
started and stopped through a web page.
The configuration concerns:
 - the SSID,
 - the WiFi channel,
 - the number of stations,
 - the security (WPA, passphrase/PSK, visibility...)
 - the IP adresses range (with check of the input)
Logs are available as well in order to know the AP status.
It includes a correction concerning the shutdown of the hostapd daemon.
WARNING: The last line with the conf-file option in the dnsmasq.conf
configuration file has to be removed. Otherwise, the dnsmasq daemon will
not start properly and as a result, the AP will not work.

Change-Id: I74e9a3b53f7c802198c02acd4d50c1565dc35366
Resolved: #4684