Prevent updating WIFI clients counter when start/stop APIs fail
A global counter in le_wifiClient keeps track of the number of clients 
currently using the service. The counter is updated when 
le_wifiClient_Start() and le_wifiClient_Stop() are called. 
This way, only the first start call starts the WiFi module and only
the last stop call shutsdown the WiFi module. 

This commit enhances the way the counter is incremented by preventing
any update if one of these APIs fail.

Resolves: LE-9174
Change-Id: I0042e8e31d3f3c3aac198ac3d0fedbacf14171a5