Fix WiFi for secure networks
Currently the WiFi service is appending the networking info to a wpa_supplicant
file /etc/wpa_supplicant.conf which contains an existing network section for
unsecure networks. This is breaking connecting to a secure AP.

This was resolved by always re-creating the /tmp/wpa_supplicant file from
scratch and writing the unsecure network section from the PA adaptor code as
well.

Resolves: LE-9561
Change-Id: Ifc444c072d6105edccda53bc3e98c17000374329