Fix string handling warnings from GCC 8.2.
* apps/sample/wifiWebAp/wifiWebApComponent/wifiWebAp.c
(WifiEventHandler, StartWebServer): Here, the issue is that the buffer
used for formatting a time is of the same size (BUF_SIZE) as a buffer
which receives that time via a larger format string. GCC informs us that
we're trying to stick a string that might be to BUF_SIZE-1 characters
long into a formatted content that is itself confined to BUF_SIZE.
Since the time is nowhere near BUF_SIZE, the fix for this is to just use
a small buffer for the time string (that is HH:MM:SS).

* service/platformAdaptor/ti/pa_wifi_ap_ti.c (pa_wifiap_SetPassPhrase):
Replace strncpy and null termination steps with le_utf8_Copy based on
the destination buffer size.

Resolves: LE-12273
Change-Id: I616a43fb5654e1ab997999ea99c944e1fd8ecade