Correct errno check on strtol call
According to errno man page:
   errno
       The value in errno is significant only when the return value of
       the call indicated an error (i.e., -1 from most system calls; -1
       or NULL from most library functions); a function that succeeds is
       allowed to change errno.  The value of errno is never set to zero
       by any system call or library function.

Hence do not assume errno to be zero at the end of a successful system
call, unless errno is set explicitely to zero before invoking the system
call.

The strtol() function sets errno to ERANGE in case of an error. Set
errno to zero before calling strtol() and check against ERANGE on
return, to detect underflow or overflow errors due to bad arguments.

Resolves: LE-16338
Change-Id: Id35b8c2f46220ac4b6f3948ae95396b823271619
Signed-off-by: Anju Kunjumon <akunjumon@sierrawireless.com>